Created back button component; added back buttons to stable pools
